Move redirect code into sub package
