+++ go.mod
@@ -2,6 +2,8 @@ module github.com/fermyon/spin-redirect
 
 go 1.24
 
-require github.com/fermyon/spin/sdk/go v1.4.2
+require github.com/spinframework/spin-go-sdk v0.0.0-20250310233821-2665ad7e30a4
 
 require github.com/julienschmidt/httprouter v1.3.0 // indirect
+
+replace github.com/spinframework/spin-go-sdk => github.com/fermyon/spin-go-sdk v0.0.0-20250310233821-2665ad7e30a4
